LPD-44159 Simplify turn specific serializers into dedicated methods. Have just one bigger serializer for custom datasets and another one for system datasets
