+++ modules/apps/frontend-data-set/frontend-data-set-api/src/main/resources/com/liferay/frontend/data/set/packageinfo
@@ -1,1 +1,1 @@
-version 2.1.0
+version 2.0.0
