+++ modules/apps/frontend-data-set/frontend-data-set-api/src/main/resources/com/liferay/frontend/data/set/serializer/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
