--- modules/apps/frontend-data-set/frontend-data-set-impl/src/main/java/com/liferay/frontend/data/set/internal/action/BaseCreationMenuFDSSerializer.java
@@ -1,23 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.frontend.data.set.internal.action;
-
-import com.liferay.frontend.data.set.FDSSerializer;
-import com.liferay.frontend.data.set.internal.BaseFDSSerializer;
-import com.liferay.frontend.taglib.clay.servlet.taglib.util.CreationMenu;
-
-/**
- * @author Daniel Sanz
- */
-public abstract class BaseCreationMenuFDSSerializer
-	extends BaseFDSSerializer implements FDSSerializer<CreationMenu> {
-
-	@Override
-	public String getKey() {
-		return "creationMenu";
-	}
-
-}
