--- modules/apps/frontend-data-set/frontend-data-set-impl/src/main/java/com/liferay/frontend/data/set/internal/action/CustomBulkActionsFDSSerializerImpl.java
@@ -1,37 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.frontend.data.set.internal.action;
-
-import com.liferay.frontend.data.set.FDSSerializer;
-import com.liferay.frontend.data.set.model.FDSActionDropdownItem;
-
-import java.util.Collections;
-import java.util.List;
-
-import javax.servlet.http.HttpServletRequest;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Daniel Sanz
- */
-@Component(
-	property = "frontend.data.set.serializer.type=" + FDSSerializer.TYPE_CUSTOM,
-	service = FDSSerializer.class
-)
-public class CustomBulkActionsFDSSerializerImpl
-	extends BaseBulkActionsFDSSerializer {
-
-	@Override
-	public List<FDSActionDropdownItem> serialize(
-		String fdsName, HttpServletRequest httpServletRequest) {
-
-		// TODO
-
-		return Collections.emptyList();
-	}
-
-}
